Added eventGridProperty and spanGridProperty components, updated grid column types, added parsing to gridProperty components
